+++ examples/expo-example/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @knocklabs/expo-example
 
+## 0.1.35-canary.2
+
+### Patch Changes
+
+- @knocklabs/expo@0.3.22-canary.2
+
+## 0.1.35-canary.1
+
+### Patch Changes
+
+- @knocklabs/expo@0.3.22-canary.1
+
+## 0.1.35-canary.0
+
+### Patch Changes
+
+- @knocklabs/expo@0.3.22-canary.0
+
 ## 0.1.34
 
 ### Patch Changes