+++ examples/guide-example/CHANGELOG.md
@@ -1,5 +1,29 @@
 # guide-example
 
+## 0.0.17-canary.2
+
+### Patch Changes
+
+- Updated dependencies [e69da7b]
+  - @knocklabs/client@0.14.10-canary.2
+  - @knocklabs/react@0.7.16-canary.2
+
+## 0.0.17-canary.1
+
+### Patch Changes
+
+- Updated dependencies [c76b2d9]
+  - @knocklabs/client@0.14.10-canary.1
+  - @knocklabs/react@0.7.16-canary.1
+
+## 0.0.17-canary.0
+
+### Patch Changes
+
+- Updated dependencies [86a72cc]
+  - @knocklabs/client@0.14.10-canary.0
+  - @knocklabs/react@0.7.16-canary.0
+
 ## 0.0.16
 
 ### Patch Changes