+++ examples/nextjs-app-dir-example/CHANGELOG.md
@@ -1,5 +1,23 @@
 # nextjs-app-dir-example
 
+## 0.0.3-canary.2
+
+### Patch Changes
+
+- @knocklabs/react@0.7.16-canary.2
+
+## 0.0.3-canary.1
+
+### Patch Changes
+
+- @knocklabs/react@0.7.16-canary.1
+
+## 0.0.3-canary.0
+
+### Patch Changes
+
+- @knocklabs/react@0.7.16-canary.0
+
 ## 0.0.2
 
 ### Patch Changes