+++ examples/slack-kit-example/CHANGELOG.md
@@ -1,5 +1,23 @@
 # slack-kit-example
 
+## 0.4.1-canary.2
+
+### Patch Changes
+
+- @knocklabs/react@0.7.16-canary.2
+
+## 0.4.1-canary.1
+
+### Patch Changes
+
+- @knocklabs/react@0.7.16-canary.1
+
+## 0.4.1-canary.0
+
+### Patch Changes
+
+- @knocklabs/react@0.7.16-canary.0
+
 ## 0.4.0
 
 ### Minor Changes