+++ packages/client/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 0.14.10-canary.2
+
+### Patch Changes
+
+- e69da7b: feat: Migrates the internal store library from zustand to @tanstack/store. This is a non-breaking change that maintains backwards compatibility with the @knocklabs/client and @knocklabs/react-core packages.
+
+## 0.14.10-canary.1
+
+### Patch Changes
+
+- c76b2d9: feat: Migrates the internal store library from zustand to @tanstack/store. This is a non-breaking change that maintains backwards compatibility with the @knocklabs/client and @knocklabs/react-core packages.
+
+## 0.14.10-canary.0
+
+### Patch Changes
+
+- 86a72cc: feat: Migrates the internal store library from zustand to @tanstack/store. This is a non-breaking change that maintains backwards compatibility with the @knocklabs/client and @knocklabs/react-core packages.
+
 ## 0.14.9
 
 ### Patch Changes