+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 0.6.22-canary.2
+
+### Patch Changes
+
+- Updated dependencies [e69da7b]
+  - @knocklabs/react-core@0.6.14-canary.2
+  - @knocklabs/client@0.14.10-canary.2
+
+## 0.6.22-canary.1
+
+### Patch Changes
+
+- Updated dependencies [c76b2d9]
+  - @knocklabs/react-core@0.6.14-canary.1
+  - @knocklabs/client@0.14.10-canary.1
+
+## 0.6.22-canary.0
+
+### Patch Changes
+
+- Updated dependencies [86a72cc]
+  - @knocklabs/react-core@0.6.14-canary.0
+  - @knocklabs/client@0.14.10-canary.0
+
 ## 0.6.21
 
 ### Patch Changes